lambda/requirements.txt
pdf2image==1.16.3
Pillow==9.5.0